Update authentication to make better use of Shibboleth.
This adds a custom credential, store, and user class for
authentication. The credential checks the environment for a configured
username variable and asks the store to create a corresponding user
object.

The user object contains some wrappers around the environment variables for use throughout the application, such as a display name and the username for use in LDAP proxy authentication.